+++ github_events_analysis/src/classes/events.py
