+++ github_events_analysis/src/user_aggregation/user_aggs.py
@@ -1,22 +1,22 @@
-"""This script gathers all the functions needed to extract the repository
+"""This script gathers all the functions needed to extract the user
 aggregation metrics"""
 from pyspark.sql.dataframe import DataFrame
 
-from github_events_analysis.classes.events import Event
-from github_events_analysis.utils.events import filter_by_event_type
+from github_events_analysis.src.classes.events import Event
+from github_events_analysis.src.utils.events import filter_by_event_type
 
 
-def get_repo_aggregations(
+def get_user_aggregations(
     data: DataFrame,
 ) -> DataFrame:
     """Main function in this flow, it creates the dataset of metrics for
-    repository aggregation
+    user aggregation
 
     Args:
        data (DataFrame): Raw dataset from which extract the metrics
 
     Return:
-        metrics (DataFrame): Metrics related to repository aggregation
+        metrics (DataFrame): Metrics related to user aggregation
 
     """
     filtered_data = filter_by_event_type(
@@ -24,33 +24,32 @@ def get_repo_aggregations(
         events_to_keep=[Event.Issues, Event.PullRequest]
     )
 
-    metrics = _get_repo_metrics(
+    metrics = _get_user_metrics(
         dataset=filtered_data
     )
 
     return metrics
 
 
-def _get_repo_metrics(
+def _get_user_metrics(
     dataset: DataFrame,
 ) -> DataFrame:
-    """Get repo aggregation metrics: number of starred projects, created issues
+    """Get user aggregation metrics: number of starred projects, created issues
     and created PRs per user-date
 
     Args:
         dataset (DataFrame): Dataset from where we will get the metrics
 
     Return:
-        metrics (DataFrame): Aggregated dataset with number of users that
-            starred it, forked it and number of created issues and PRs
-
+        metrics (DataFrame): Aggregated dataset with number of
+            starred projects, created issues and created PRs
     """
     metrics = (
         dataset
         .groupBy(
             [
-                "project_id",
-                "project_name",
+                "actor_login",
+                "actor_id",
                 "day",
                 "type"
             ]
